requirements.txt
python-telegram-bot==22.5
groq==0.32.0
python-dotenv==1.1.1